requirements.txt
streamlit
Pillow
moviepy
google-generativeai
